fix: Ensure sourcemap and substitution integrity
Previously, nutra-mock would manipulate the source, but not
generate a source map. This broke coverage as the original source
did not match the transpiler's sourcemap. In addition to that,
constant substitutions was fairly unreliable. It targeted
everything on the page, with the potential to really screw things
up. This commit parses and traverses the entire ast tree and makes
decisions based on node's details. It also replaces nodes with
other nodes, making it possible to automatically generate a source
map. Essentially killing two birds with one stone. This however,
does not solve the problem of deep context constants with the same
name as a global constant. This could be solved with a scope
analysis, which might not be trivial.