fix: Ensure deep context constants with repeated names escape
Previously, if there a constant in the main execution context
called "Foo" and "Foo" was declared again in a child context, it
would overwrite "Foo" without realizing that it might be something
entirely different. This commit fixes the issue by keeping track
and excluding modifications to local variables upon entering the
scope and restoring them when leaving.