fix: Ensure get mock method returns the mock
Previously, NutraMock.getEntry('bar').get method was always
returning undefined. This occured simply because it did not return
the value, although it did have it. This commit removes the
brackets around the function -- ensuring that it returns the value
from the method it calls, which is the mock value we want.